Fix for issue #329
Split examples in multiple files
